+++ includes/vector/op_vec.h
@@ -19,9 +19,9 @@ t_vec3	min_vec(t_vec3 a, t_vec3 b);
 
 t_vec3	plus_vec(t_vec3 a, t_vec3 b);
 
-double	dot(t_vec3 a, t_vec3 b);
+float	dot(t_vec3 a, t_vec3 b);
 
-double	len3(t_vec3 a);
+float	len3(t_vec3 a);
 
 t_vec3	normalize(t_vec3 a);
