+++ src/intersections/plane.c
@@ -6,15 +6,15 @@
 #include <op_vec.h>
 #include <math.h>
 
-double hit_plane(t_plane *plane, t_ray r)
+float hit_plane(t_plane *plane, t_ray r)
 {
-	double dn_dot;
-	double t;
+	float dn_dot;
+	float t;
 	t_vec3 normal;
 	t_vec3 tmp;
 	normal = (plane->orientation);
 	dn_dot = dot(r.direction, normal);
-	if (fabs(dn_dot) > 1e-6)
+	if (fabsf(dn_dot) > 1e-6)
 	{
 		tmp = min_vec(r.direction, normal);
 		t = dot(tmp, normal) / dn_dot;
