+++ src/rendering/colors.c
@@ -12,9 +12,9 @@
 
 
 
-t_list *ray_color(t_ray r, t_list **head, double *distance)
+t_list *ray_color(t_ray r, t_list **head, float *distance)
 {
-	double t;
+	float t;
 	t_list *elem;
 	t_list *hit_elem;
 
@@ -40,7 +40,7 @@ t_list *ray_color(t_ray r, t_list **head, double *distance)
 	return (hit_elem);
 }
 
-t_colour create_obj(t_list *hit_elem, t_ray r, t_data img, double distance, t_list **head)
+t_colour create_obj(t_list *hit_elem, t_ray r, t_data img, float distance, t_list **head)
 {
 	t_object obj;
