+++ src/rendering/colors_utils.c
@@ -8,7 +8,7 @@
 #include <op_vec.h>
 #include <math.h>
 
-int	clip_colour(double c)
+int	clip_colour(float c)
 {
 	if (c < 0)
 		return (0);
@@ -38,6 +38,6 @@ t_colour	get_ray_luminosity(t_data img, t_object obj, t_ray r)
 	if (img.light.brightness == 0)
 		return (c);
 	c = plus_vec(c, mult3(obj.colour, img.light.brightness * dot(r.direction, obj.normal_to_surface)));
-	c = plus_vec(c, mult3(specular, pow(img.light.brightness * dot(obj.normal_to_surface, normalize(plus_vec(r.direction, min_vec(img.camera.view_point, obj.intersection)))), 4 * len3(min_vec(img.light.coordinates, obj.intersection)))));
+	c = plus_vec(c, mult3(specular, powf(img.light.brightness * dot(obj.normal_to_surface, normalize(plus_vec(r.direction, min_vec(img.camera.view_point, obj.intersection)))), 4 * len3(min_vec(img.light.coordinates, obj.intersection)))));
 	return (c);
 }
